KAF-39: Partitions don't rebalance in distributed mode across multiple workers
Only one worker in a connector group starts a DseSinkConnector
instance. Common startup logic is now refactored to SinkUtil and
set up to be invoked by DseSinkTask.start such that only one task
actually performs the startup logic and other tasks wait for the
result.

Similarly, when a task is stopped, the session is closed by the last
task in the connector instance.